arch-x86: Partial AVX-512 Support.
This patch initializes some partial support for AVX-512.
It is not in submittable state, but should be sufficient
to solicit some feedbacks as the first step.

What's included:
1. Extended xmm to zmm registers. Added mask k0-k7 regs.
2. Extended the decoder to parse EVEX prefix.
3. A base class AVXOpBase for all AVX micro ops.
4. Implementation of common/frequently used instructions.

Known issues:
1. Very limited mask support, i.e. only a few instructions
like vcmpf handles masks.
2. Not throughly tested -- I just implement new
instructions encountered during my research.
3. Due to some legacy code, this implementation still
assumes a fixed array in `AVXOpBase` for operand and
destination register ids, while other x86 microops
customize the size of this array.
4. Some instructions are still implemented using the
`mediaop`, but this can be fixed.

What's the plan:
I hope to get some feedbacks on the high-level design
and whether we are interested to push this forward.
Since it's a big feature taking I would say a reasonable
plan would be create a separate branch where I can push
the basic skeleton code. And I and other
contributors (hopefully there are) can write the test
cases and push new instructions.

Change-Id: I9f8c53043f181c2f818200558b7cdef7436c173d